requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
deepgram-sdk
elevenlabs
openai
anthropic
twilio
neo4j
lyzr-agent-api
python-multipart
httpx
motor
websockets
python-jose[cryptography]
passlib[bcrypt]
